finished cleaning data, adding overall data
This push contains a lot because I had to split the data into separate dataframes (each distance of radii gets its own dataframe, and each distance of radii per hq gets its own dataframe). because there is so much code to clean the base and dissagregated data, I've made separate clean_data files. this will need to be cleaned up later, but is fine for dissertation