add code for all models (clean/analyze/plot)
This is a large commit that will need to be cleaned up later. for now, cleaning the data overall and by hq is separated, as is the the analyzing. plotting the data is relatively easy, so is completed in one script. Later on, will organize the scripts and data outputs better, but this will suffice for now